fixed 'multiple organisations' options not showing due to race conditions with loggedInUser state
