+++ src/index.tsx
@@ -12,15 +12,18 @@ import { BrowserRouter } from "react-router-dom";
 import reportWebVitals from "./reportWebVitals";
 import { QueryClientProvider } from "react-query";
 import { queryClient } from "./utils/queryClient";
+import { OrganisationProvider } from "./context/OrganisationContext";
 
 ReactDOM.render(
   <React.StrictMode>
-    <QueryClientProvider client={queryClient}>
-      <BrowserRouter>
-        <App />
-      </BrowserRouter>
-      {/* <ReactQueryDevtools initialIsOpen={false} /> */}
-    </QueryClientProvider>
+    <OrganisationProvider>
+      <QueryClientProvider client={queryClient}>
+        <BrowserRouter>
+          <App />
+        </BrowserRouter>
+        {/* <ReactQueryDevtools initialIsOpen={false} /> */}
+      </QueryClientProvider>
+    </OrganisationProvider>
   </React.StrictMode>,
   document.getElementById("root")
 );
